Revert "update tooling, no longer use babel-node in production"
This reverts commit 779c80bebed59f659ba096be0a1433cce3bd6978.
