+++ package.json
@@ -5,12 +5,12 @@
   "main": "server.js",
   "scripts": {
     "coverage": "babel-node node_modules/isparta/bin/isparta cover --report lcov node_modules/mocha/bin/_mocha -- --reporter dot",
-    "dist": "webpack -p --config webpack.config.js",
+    "dist": "webpack --config webpack.config.js",
     "build": "npm run clean && npm run dist",
-    "clean": "rm -rf dist && rm -rf server.bundle.js",
+    "clean": "rm -rf dist",
     "start": "npm run start:dev",
-    "start:dev": "NODE_ENV=development babel-node serverDev.js",
-    "start:prod": "NODE_ENV=production PORT=5000 node server.bundle.js",
+    "start:dev": "NODE_ENV=development babel-node server.js",
+    "start:prod": "NODE_ENV=production babel-node server.js",
     "test": "mocha --compilers js:babel-core/register --recursive && npm run coverage",
     "test:watch": "npm test -- --watch",
     "lint": "eslint src"
@@ -43,7 +43,6 @@
     "clipboard": "^1.5.5",
     "codeclimate-test-reporter": "^0.3.0",
     "codemirror": "5.11.0",
-    "compression": "1.6.1",
     "cookie-parser": "^1.4.1",
     "es6-promise": "^3.0.2",
     "eslint": "^1.10.3",
