+++ server.js
@@ -1,12 +1,16 @@
 import path from 'path';
 import express from 'express';
+import webpack from 'webpack';
+import webpackMiddleware from 'webpack-dev-middleware';
+import webpackHotMiddleware from 'webpack-hot-middleware';
+import config from './webpack.config.dev';
 import cookieParser from 'cookie-parser';
 import helmet from 'helmet';
 import bodyParser from 'body-parser';
 import morgan from 'morgan';
-import compression from 'compression';
 
-const port = process.env.PORT || 5000;
+const isDeveloping = process.env.NODE_ENV !== 'production';
+const port = isDeveloping ? 5000 : process.env.PORT;
 const server = global.server = express();
 
 server.disable('x-powered-by');
@@ -15,13 +19,34 @@ server.use(helmet());
 server.use(bodyParser.urlencoded({ extended: false }));
 server.use(bodyParser.json());
 server.use(cookieParser());
-server.use(morgan('combined'));
-server.use(compression());
+server.use(morgan('dev'));
 
-server.use('/static', express.static(__dirname + '/dist'));
-server.get('*', function response(req, res) {
-  res.sendFile(path.join(__dirname, '/index.html'));
-});
+if (isDeveloping) {
+  const compiler = webpack(config);
+  const middleware = webpackMiddleware(compiler, {
+    publicPath: config.output.publicPath,
+    contentBase: 'src',
+    stats: {
+      colors: true,
+      hash: false,
+      timings: true,
+      chunks: false,
+      chunkModules: false,
+      modules: false,
+    },
+  });
+
+  server.use(middleware);
+  server.use(webpackHotMiddleware(compiler));
+  server.get('*', function response(req, res) {
+    res.sendFile(path.join(__dirname, '/index.html'));
+  });
+} else {
+  server.use('/static', express.static(__dirname + '/dist'));
+  server.get('*', function response(req, res) {
+    res.sendFile(path.join(__dirname, '/index.html'));
+  });
+}
 
 server.use('/api/v0/extract', require('./api/extract'));
 server.use('/api/v0/premail', require('./api/premail'));
