+++ src/store/configureStore.js
@@ -16,13 +16,10 @@ export default function configureStore(initialState) {
     )
   );
 
-  if (__DEV__) {
-    console.log('in development');
-    if (module.hot) {
-      module.hot.accept('../reducers', () =>
-        store.replaceReducer(require('../reducers').default /*.default if you use Babel 6+ */)
-      );
-    }
+  if (module.hot) {
+    module.hot.accept('../reducers', () =>
+      store.replaceReducer(require('../reducers').default /*.default if you use Babel 6+ */)
+    );
   }
 
   return store;
