+++ webpack.config.dev.js
@@ -18,9 +18,8 @@ module.exports = {
     new webpack.optimize.OccurenceOrderPlugin(),
     new webpack.HotModuleReplacementPlugin(),
     new webpack.NoErrorsPlugin(),
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify('development'),
-      __DEV__: true
+     new webpack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify('development')
     }),
   ],
   module: {
